Fix break route for CykelPlanen

+++ I Bike CPH/CykelPlanen/Classes/Views/SMAddressPickerView.h
@@ -30,8 +30,8 @@ typedef enum  {
 @property (weak, nonatomic) IBOutlet UIPickerView *pickerView;
 @property (assign, nonatomic) AddressType addressType;
 @property (weak, nonatomic) id<AddressSelectDelegate> delegate;
-@property (assign, nonatomic) int sourceCurrentIndex;
-@property (assign, nonatomic) int destinationCurrentIndex;
+@property (assign, nonatomic) NSInteger sourceCurrentIndex;
+@property (assign, nonatomic) NSInteger destinationCurrentIndex;
 
 - (IBAction)didTapOnCancel:(id)sender;
 - (IBAction)didTapOnDone:(id)sender;
